IN-1112 fix count errors

+++ migrate.sh
@@ -89,21 +89,24 @@ fi
 
 docker-compose ${COMPOSE_ARGS} run --rm initialise initialise_environments/initialise_environments.sh -i "${PRESERVE_SCHEMAS}"
 
-docker rm casrec_load_1 &>/dev/null || echo "casrec_load_1 does not exist. This is OK"
-docker rm casrec_load_2 &>/dev/null || echo "casrec_load_2 does not exist. This is OK"
-docker rm casrec_load_3 &>/dev/null || echo "casrec_load_3 does not exist. This is OK"
-docker rm casrec_load_4 &>/dev/null || echo "casrec_load_4 does not exist. This is OK"
-docker-compose ${COMPOSE_ARGS} run --rm --name casrec_load_1 load_casrec python3 load_casrec/app/app.py -p "1" -t "4" >> docker_load.log &
-P1=$!
-docker-compose ${COMPOSE_ARGS} run --rm --name casrec_load_2 load_casrec python3 load_casrec/app/app.py -p "2" -t "4" >> docker_load.log &
-P2=$!
-docker-compose ${COMPOSE_ARGS} run --rm --name casrec_load_3 load_casrec python3 load_casrec/app/app.py -p "3" -t "4" >> docker_load.log &
-P3=$!
-docker-compose ${COMPOSE_ARGS} run --rm --name casrec_load_4 load_casrec python3 load_casrec/app/app.py -p "4" -t "4" >> docker_load.log &
-P4=$!
-wait $P1 $P2 $P3 $P4
-cat docker_load.log
-rm docker_load.log
+if [ "${REBUILD_CASREC_CSV_SCHEMA}" == "y" ]
+then
+  docker rm casrec_load_1 &>/dev/null || echo "casrec_load_1 does not exist. This is OK"
+  docker rm casrec_load_2 &>/dev/null || echo "casrec_load_2 does not exist. This is OK"
+  docker rm casrec_load_3 &>/dev/null || echo "casrec_load_3 does not exist. This is OK"
+  docker rm casrec_load_4 &>/dev/null || echo "casrec_load_4 does not exist. This is OK"
+  docker-compose ${COMPOSE_ARGS} run --rm --name casrec_load_1 load_casrec python3 load_casrec/app/app.py -p "1" -t "4" >> docker_load.log &
+  P1=$!
+  docker-compose ${COMPOSE_ARGS} run --rm --name casrec_load_2 load_casrec python3 load_casrec/app/app.py -p "2" -t "4" >> docker_load.log &
+  P2=$!
+  docker-compose ${COMPOSE_ARGS} run --rm --name casrec_load_3 load_casrec python3 load_casrec/app/app.py -p "3" -t "4" >> docker_load.log &
+  P3=$!
+  docker-compose ${COMPOSE_ARGS} run --rm --name casrec_load_4 load_casrec python3 load_casrec/app/app.py -p "4" -t "4" >> docker_load.log &
+  P4=$!
+  wait $P1 $P2 $P3 $P4
+  cat docker_load.log
+  rm docker_load.log
+fi
 echo "=== Step 0 - Filter data ==="
 docker-compose ${COMPOSE_ARGS} run --rm initialise prepare_source_data/prepare_source_data.sh
 echo "=== Step 1 - Transform ==="
